+++ lib/active_storage_validations/matchers/content_type_validator_matcher.rb
@@ -10,6 +10,7 @@ require_relative "shared/asv_contextable"
 require_relative "shared/asv_except_onable"
 require_relative "shared/asv_messageable"
 require_relative "shared/asv_rspecable"
+require_relative "shared/asv_spoofing_protectable"
 require_relative "shared/asv_timeoutable"
 require_relative "shared/asv_validatable"
 
@@ -27,6 +28,7 @@ module ActiveStorageValidations
       include ASVExceptOnable
       include ASVMessageable
       include ASVRspecable
+      include ASVSpoofingProtectable
       include ASVTimeoutable
       include ASVValidatable
 
@@ -36,6 +38,7 @@ module ActiveStorageValidations
         initialize_except_onable
         initialize_messageable
         initialize_rspecable
+        initialize_spoofing_protectable
         initialize_timeoutable
         @attribute_name = attribute_name
         @allowed_content_types = @rejected_content_types = []
@@ -71,6 +74,7 @@ module ActiveStorageValidations
           is_except_on_valid? &&
           is_allowing_blank? &&
           is_timeout_valid? &&
+          is_spoofing_protection_valid? &&
           is_custom_message_valid? &&
           all_allowed_content_types_allowed? &&
           all_rejected_content_types_rejected?